Moved all allUTXOsToFedAreAboveMinimumPeginValue tests to a new class called PegUtilsAllUTXOsToFedAreAboveMinimumPeginValueTest and added more tests for the retiring fed
