+++ rskj-core/src/test/java/co/rsk/peg/PegUtilsTest.java
@@ -1,6 +1,12 @@
 package co.rsk.peg;
 
-import co.rsk.bitcoinj.core.*;
+import co.rsk.bitcoinj.core.BtcTransaction;
+import co.rsk.bitcoinj.core.Coin;
+import co.rsk.bitcoinj.core.Context;
+import co.rsk.bitcoinj.core.NetworkParameters;
+import co.rsk.bitcoinj.core.Address;
+import co.rsk.bitcoinj.core.BtcECKey;
+import co.rsk.bitcoinj.core.Sha256Hash;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.bitcoinj.script.ScriptBuilder;
 import co.rsk.bitcoinj.wallet.Wallet;
@@ -630,329 +636,6 @@ class PegUtilsTest {
 
         // Assert
         Assertions.assertEquals(PegTxType.UNKNOWN, pegTxType);
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PreRSKIP379AndNoOutputs_true() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForFingerrot = ActivationConfigsForTest.fingerroot500().forBlock(0);
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertTrue(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForFingerrot)
-                , "When RSKIP379 is not yet active and the utxo list to the fed is empty, it returns true."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PostRSKIP379AndNoOutputs_false() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForTbd600 = ActivationConfigsForTest.tbd600().forBlock(0);
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertFalse(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForTbd600)
-                , "When RSKIP379 is active and the utxo list to the fed is empty, it should return false."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PreRSKIP379WithManyOutputsBelowMinimumButNoneForFed_true() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForFingerrot = ActivationConfigsForTest.fingerroot500().forBlock(0);
-
-        Address randomAddress = BitcoinTestUtils.createP2PKHAddress(btcMainnetParams, "add1");
-
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress);
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress);
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress);
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertTrue(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForFingerrot)
-                , "When RSKIP379 is not yet active and there are many utxos but none for the fed, it returns true."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PostRSKIP379WithManyOutputsBelowMinimumButNoneForFed_false() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForTbd600 = ActivationConfigsForTest.tbd600().forBlock(0);
-
-        Address randomAddress = BitcoinTestUtils.createP2PKHAddress(btcMainnetParams, "add1");
-
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress);
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress);
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress);
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertFalse(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForTbd600)
-                , "When RSKIP379 is active and there are many utxos but none for the fed, it should return false."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PreRSKIP379WithManyOutputsEqualAndAboveMinimumButNoneForFed_true() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForFingerrot = ActivationConfigsForTest.fingerroot500().forBlock(0);
-
-        Address randomAddress = BitcoinTestUtils.createP2PKHAddress(btcMainnetParams, "add1");
-
-        btcTx.addOutput(minimumPegInTxValue, randomAddress); // Minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), randomAddress); // Above minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), randomAddress); // Above minimum
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertTrue(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForFingerrot)
-                , "When RSKIP379 is not yet active and there are many utxos equal and above minimum but none for the fed, it returns true."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PostRSKIP379WithManyOutputsEqualAndAboveMinimumButNoneForFed_false() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForTbd600 = ActivationConfigsForTest.tbd600().forBlock(0);
-
-        Address randomAddress = BitcoinTestUtils.createP2PKHAddress(btcMainnetParams, "add1");
-
-        btcTx.addOutput(minimumPegInTxValue, randomAddress); // Minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), randomAddress); // Above minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), randomAddress); // Above minimum
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertFalse(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForTbd600)
-                , "When RSKIP379 is active and there are many utxos equal and above minimum but none for the fed, it returns false."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PreRSKIP379WithManyOutputsAboveMinimumForFedAndManyOutputsBelowMinimumForRandomAddresses_true() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForFingerrot = ActivationConfigsForTest.fingerroot500().forBlock(0);
-
-        Address randomAddress = BitcoinTestUtils.createP2PKHAddress(btcMainnetParams, "add1");
-
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress); // Below Minimum
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress); // Below Minimum
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress); // Below Minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertTrue(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForFingerrot)
-                , "When RSKIP379 is not yet active and there are many utxos for fed above minimum and many utxos to random addresses below minimum, it returns true."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PostRSKIP379WithManyOutputsAboveMinimumForFedAndManyOutputsBelowMinimumForRandomAddresses_true() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForTbd600 = ActivationConfigsForTest.tbd600().forBlock(0);
-
-        Address randomAddress = BitcoinTestUtils.createP2PKHAddress(btcMainnetParams, "add1");
-
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress); // Below Minimum
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress); // Below Minimum
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), randomAddress); // Below Minimum
-        btcTx.addOutput(minimumPegInTxValue, activeFederation.getAddress()); // Minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertTrue(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForTbd600)
-                , "When RSKIP379 is active and there are many utxos for fed above minimum and many utxos to random addresses below minimum, it returns true."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PreRSKIP379WithManyOutputsForFedButEachBelowMinimumAndSumAboveMinimum_false() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForFingerrot = ActivationConfigsForTest.fingerroot500().forBlock(0);
-
-        Coin amountJustBelowMinimum = minimumPegInTxValue.minus(Coin.SATOSHI);
-
-        btcTx.addOutput(amountJustBelowMinimum, activeFederation.getAddress());
-        btcTx.addOutput(amountJustBelowMinimum, activeFederation.getAddress());
-        btcTx.addOutput(amountJustBelowMinimum, activeFederation.getAddress());
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertFalse(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForFingerrot)
-                , "When RSKIP379 is not yet active and there are many utxos to the fed but there is at least one of them below the minimum, it returns false."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PostRSKIP379WithManyOutputsForFedButEachBelowMinimumAndSumAboveMinimum_false() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForTbd600 = ActivationConfigsForTest.tbd600().forBlock(0);
-
-        Coin amountJustBelowMinimum = minimumPegInTxValue.minus(Coin.SATOSHI);
-
-        btcTx.addOutput(amountJustBelowMinimum, activeFederation.getAddress());
-        btcTx.addOutput(amountJustBelowMinimum, activeFederation.getAddress());
-        btcTx.addOutput(amountJustBelowMinimum, activeFederation.getAddress());
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertFalse(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForTbd600)
-                , "When RSKIP379 is active and there are many utxos to the fed but there is at least one of them below the minimum," +
-                        " even though the sum of all those utxos is above minimum, it should return false."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PreRSKIP379WithManyOutputsForFedAboveMinimumAndOneBelowMinimum_false() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForFingerrot = ActivationConfigsForTest.fingerroot500().forBlock(0);
-
-        btcTx.addOutput(minimumPegInTxValue, activeFederation.getAddress()); // With minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), activeFederation.getAddress()); // Below minimum
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertFalse(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForFingerrot)
-                , "When RSKIP379 is not yet active and there are many utxos to the fed but there is at least one of them below the minimum, it returns false."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PostRSKIP379WithManyOutputsForFedAboveMinimumAndOneBelowMinimum_false() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForTbd600 = ActivationConfigsForTest.tbd600().forBlock(0);
-
-        btcTx.addOutput(minimumPegInTxValue, activeFederation.getAddress()); // With minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-        btcTx.addOutput(minimumPegInTxValue.minus(Coin.SATOSHI), activeFederation.getAddress()); // Below minimum
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertFalse(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForTbd600)
-                , "When RSKIP379 is active and there are many utxos to the fed but there is at least one of them below the minimum, it should return true."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PreRSKIP379WithManyOutputsForFedAllEqualOrAboveMinimum_true() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForFingerroot = ActivationConfigsForTest.fingerroot500().forBlock(0);
-
-        btcTx.addOutput(minimumPegInTxValue, activeFederation.getAddress()); // With minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertTrue(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForFingerroot)
-                , "When RSKIP379 is not yet active and there are many utxos to the fed above minimum, it returns true."
-        );
-
-    }
-
-    @Test
-    void test_allUTXOsToFedAreAboveMinimumPeginValue_PostRSKIP379WithManyOutputsForFedAllEqualOrAboveMinimum_true() {
-
-        Coin minimumPegInTxValue = bridgeMainnetConstants.getMinimumPeginTxValue(activations);
-
-        BtcTransaction btcTx = new BtcTransaction(bridgeMainnetConstants.getBtcParams());
-
-        ActivationConfig.ForBlock activationsForTbd600 = ActivationConfigsForTest.tbd600().forBlock(0);
-
-        btcTx.addOutput(minimumPegInTxValue, activeFederation.getAddress()); // With minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-        btcTx.addOutput(minimumPegInTxValue.add(Coin.SATOSHI), activeFederation.getAddress()); // Above minimum
-
-        Wallet fedWallet = new BridgeBtcWallet(context, Collections.singletonList(activeFederation));
-
-        Assertions.assertTrue(
-                PegUtils.allUTXOsToFedAreAboveMinimumPeginValue(minimumPegInTxValue, btcTx, fedWallet, activationsForTbd600)
-                , "When RSKIP379 is active and there are many utxos to the fed above minimum, it returns true."
-        );
-
     }
 
 }
